requirements.txt
Flask
Flask-SocketIO
tensorflow
opencv-python
numpy
Pillow
jinja2
markupsafe
